requirements.txt
coverage
setuptools
numpy
pillow